Further causal loop tests.  Moved extract_loops from visualization to CausalLoop
