+++ src/CausalLoop.jl
@@ -2,9 +2,12 @@ export TheoryCausalLoop, AbstractCausalLoop, CausalLoopUntyped, CausalLoop, Caus
 nn, ne, nname,
 sedge, tedge, convertToCausalLoop, nnames, CausalLoopF, epol, epols,
 add_node!, add_nodes!, add_edge!, add_edges!, discard_zero_pol,
-outgoing_edges, incoming_edges
+outgoing_edges, incoming_edges, extract_loops
 
 using MLStyle
+
+import Graphs: SimpleDiGraph, simplecycles, SimpleEdge
+
 
 
 
@@ -163,6 +166,79 @@ function convertToCausalLoop(p::AbstractStockAndFlowStructureF)
     es=vcat(lses,lsvfes,lfves,fies,foes,lpvs,lvvs)
 
     return CausalLoop(ns,es)
+end
+
+""" 
+Cycles are uniquely characterized by sets of edges, not sets of nodes
+We construct a simple graph, then for each edge, we check if there exist
+multiple edges with the same start and end node
+We then product all of them, to get every cycle.
+
+This could be made more efficient, but it should be fine for now.
+"""
+function extract_loops(cl::CausalLoopF)
+
+  edges = collect(zip(subpart(cl, :s), subpart(cl, :t)))
+  pair_to_edge = state_dict(edges) # Note, this is unique pairs, not all.
+  # Unique are sufficient for making simple graph.
+  g = SimpleDiGraph(SimpleEdge.(edges))
+
+  # Edges -> Polarity
+  cycle_pol = Dict{Vector{Int}, Polarity}()
+  for cycle ∈ simplecycles(g)
+    cycle_length = length(cycle)
+    # Last pair is cycle[end], cycle[1]
+    node_pairs = ((cycle[node_index], cycle[(node_index % cycle_length) + 1]) for node_index in 1:cycle_length)
+    nonsimple_cycles = Vector{Vector{Int}}()
+    for (p1, p2) in node_pairs
+      matching_edges = Vector{Int}()
+      # grab all edges with p1 as start and p2 as end
+      # This is the bit that could be made more efficient, it loops over all edges every time
+      for cl_node in 1:ne(cl)
+        if sedge(cl, cl_node) == p1 && tedge(cl, cl_node) == p2
+          push!(matching_edges, cl_node)
+        end
+      end
+      push!(nonsimple_cycles, matching_edges)
+    end
+
+    for cycle_instance in Base.Iterators.product(nonsimple_cycles...)
+      balancing_count = 0
+      is_unknown = false
+      is_zero = false
+      is_not_well_defined = false
+      for node_number in cycle_instance
+        current_pol = epol(cl, node_number)
+        if current_pol == POL_BALANCING
+          balancing_count += 1
+        elseif current_pol == POL_UNKNOWN
+          is_unknown = true
+        elseif current_pol == POL_ZERO
+          is_zero = true
+          break
+        elseif current_pol == POL_NOT_WELL_DEFINED
+          is_not_well_defined = true
+        end
+      end
+
+      collected_cycle = collect(cycle_instance)
+
+      if is_zero
+        push!(cycle_pol, collected_cycle => POL_ZERO)
+      elseif is_unknown
+        push!(cycle_pol, collected_cycle => POL_UNKNOWN)
+      elseif is_not_well_defined
+        push!(cycle_pol, collected_cycle => POL_NOT_WELL_DEFINED)
+      elseif iseven(balancing_count)
+        push!(cycle_pol, collected_cycle => POL_REINFORCING)
+      else
+        push!(cycle_pol, collected_cycle => POL_BALANCING)
+      end
+    end
+  end
+
+  cycle_pol
+          
 end
